Split CDEF functions to their own source file (#371)
No need to clutter up lib.rs with self-contained CDEF work.
Split the CDEF functions into their own file.  No functional change.
